requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
